+++ libraries/PCF8574/PCF8574.h
@@ -3,13 +3,10 @@
 //    FILE: PCF8574.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 02-febr-2013
-// VERSION: 0.3.4
+// VERSION: 0.3.5
 // PURPOSE: Arduino library for PCF8574 - 8 channel I2C IO expander
 //     URL: https://github.com/RobTillaart/PCF8574
 //          http://forum.arduino.cc/index.php?topic=184800
-//
-// HISTORY:
-// see PCF8574.cpp file
 //
 
 
@@ -17,7 +14,7 @@
 #include "Wire.h"
 
 
-#define PCF8574_LIB_VERSION         (F("0.3.4"))
+#define PCF8574_LIB_VERSION         (F("0.3.5"))
 
 #ifndef PCF8574_INITIAL_VALUE
 #define PCF8574_INITIAL_VALUE       0xFF
@@ -57,11 +54,11 @@ public:
 
 
   //added 0.1.07/08 Septillion
-  inline uint8_t readButton8()  { return PCF8574::readButton8(_buttonMask); }
-  uint8_t        readButton8(const uint8_t mask);
-  uint8_t        readButton(const uint8_t pin);
-  inline void    setButtonMask(const uint8_t mask) { _buttonMask = mask; };
-  uint8_t        getButtonMask() { return _buttonMask; };
+  uint8_t readButton8()  { return PCF8574::readButton8(_buttonMask); }
+  uint8_t readButton8(const uint8_t mask);
+  uint8_t readButton(const uint8_t pin);
+  void    setButtonMask(const uint8_t mask) { _buttonMask = mask; };
+  uint8_t getButtonMask() { return _buttonMask; };
 
 
   // rotate, shift, toggle, reverse expect all lines are output
@@ -72,6 +69,12 @@ public:
   void    rotateRight(const uint8_t n = 1);
   void    rotateLeft(const uint8_t n = 1);
   void    reverse();
+
+
+  void    select(const uint8_t pin);
+  void    selectN(const uint8_t pin);
+  void    selectNone() { write8(0x00); };
+  void    selectAll()  { write8(0xFF); };
 
 
   int     lastError();